requirements.txt
openai-whisper
torch==2.6.0
torchaudio==2.6.0
streamlit
torchvision==0.21.0
huggingface-hub
ollama
langchain-community
langchain
sounddevice
torchtune
torchao
-U moshi
tokenizers
transformers==4.46.1
flask 
flask-cors 
python-dotenv